some compile fixes for gcc 4.3 and work on making more stuff compile again after the big renaming started
git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dune-mux/trunk@899 2fb0f335-1f38-0410-981e-8018bf24f1b0

+++ dumux/brinkman/fv/fvbrinkman.hh
@@ -595,7 +595,7 @@ namespace Dune
 		vtkwriter.addCellData(this->pressure,"total pressure p~");
 //		for (int i = 0; i < dim; i++)
 //		{
-			for (int k = 0; k < velocityComponentX.size(); k++)
+        for (int k = 0; k < (int) velocityComponentX.size(); k++)
 			{
 				velocityComponentX[k] = elementVelocity[k][0];
 				velocityComponentY[k] = elementVelocity[k][1];
